Numeric Casts
Try to run the following code: double x = (11+4)/2
What answer do you get? The answer is imprecise!
Write a Java program that precisely calculates the average between two integers that input from the keyboard, and prints the answer to the screen. The average must include the integer and decimal portions of the calculation.
Sample Run:

Please enter two integers:
4
11

The average is: 7.5