Processing Arrays of Strings
Due to a problem with a scanner an array of words was created with spaces in incorrect places. Write the code to process the list of words and trim any spaces out of the words.  Note that this activity does not require you to print anything. Your code should end with the array list still declared and containing the resulting words.
So if the list contains:
{"every", " near ing ", " checking", "food ", "stand", "value "}
It should be changed to hold:
{"every", "nearing", "checking", "food", "stand", "value"}