Binary & Octal
Octal, or base eight, numbers work the same way as base 2. The base is eight, and we use the digits 0 - 7. For this program you will input an int to represent the octal number and translate to the base ten number. The octal number must be 8 digits or less. Your program should also check that all the digits are 0 - 7, then translate the number to base ten.
Sample Run 1:

Enter a number in base 8:
123 
83
Sample Run 2:

Enter a number in base 8: 
1287 
ERROR: Incorrect Octal Format
Sample Run 3:

Enter a number in base 8: 
1111111111 
ERROR: Incorrect Octal Format