Parameters
Write a method that takes a parameter for the number of a month and prints the number of days in the month. Assume that February will always have 28 days for this activity. This method must be called monthDays() and it must take a integer parameter. Calling monthDays(2) would print 28 and monthDays(9) would print 30.