Parameters - Primitive vs. Class
Write a method that takes an array of ints, an integer value, and an integer index. The method should insert the value at the given index and move the values afterwards by one. This method must be called insertValue() and must have the following parameter types: int[], integer, integer. For example, insertValue(a, 100, 2)would change the array {1, 2, 3, 4, 5} to {1, 2, 100, 3, 4}.