Return Methods
Write a method that takes an array of ints as a parameter and returns the sum of integers in the array. This method must be named sum() and it must have an int[] parameter. This method must return an int. Calling sum(a) would return 6 if a = {1, 2, 3} or 3 if a = {1, 1, 1}.