Return Methods
Write a method that takes an array of ints as a parameter and returns the average value of the array as a double. This method must be named average() and it must have an int[] parameter. This method must return a double. Calling average(a) would return 2.0 if a = {1, 2, 3} or 1.0 if a = {1, 1, 1}.