Return Methods
Write a method that takes an array of ints and returns true if all of the values in the array are nonnegative (greater than or equal to zero). If the array contains any negative integers, it should return false. This method must be named allPositive() and have an int[] parameter. This method must return a boolean.